node-main/test/fixtures/wpt/encoding/legacy-mb-schinese/gbk/gbk-encoder.html

<!doctype html>
<meta charset=gbk> <!-- if the server overrides this, it is stupid, as this is a testsuite -->
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
 function encode(input, output, desc) {
   test(function() {
     const a = document.createElement("a") // <a> uses document encoding for URL's query
     a.href = "https://example.com/?" + input;
     assert_equals(a.search.substr(1), output); // remove leading "?"
   }, "gbk encoder: " + desc);
 }

 encode("s", "s", "very basic");
 encode("\u20AC", "%80", "Euro");
 encode("\u4E02", "%81@", "character");
 encode("\uE4C6", "%A1@", "PUA");
 encode("\uE4C5", "%FE%FE", "PUA #2");
 encode("\ud83d\udca9", "%26%23128169%3B", "poo");
 encode("\uE7C8", "%26%2359336%3B", "legacy ICU special case 1");
 encode("\u2026", "%A1%AD", "legacy ICU special case 2");
 encode("\uFF5E", "%A1%AB", "legacy ICU special case 3");
 encode("\u00A5", "%26%23165%3B", "legacy WebKit case 1");
 encode("\u22EF", "%26%238943%3B", "legacy WebKit case 2");
 encode("\u301C", "%26%2312316%3B", "legacy WebKit case 3");
 encode("\u{10FFFF}", "%26%231114111%3B", "U+10FFFF");

  // GB18030-2022
 encode("\uFE10", "%A6%D9", "GB18030-2022 1");
 encode("\uFE12", "%A6%DA", "GB18030-2022 2");
 encode("\uFE11", "%A6%DB", "GB18030-2022 3");
 encode("\uFE13", "%A6%DC", "GB18030-2022 4");
 encode("\uFE14", "%A6%DD", "GB18030-2022 5");
 encode("\uFE15", "%A6%DE", "GB18030-2022 6");
 encode("\uFE16", "%A6%DF", "GB18030-2022 7");
 encode("\uFE17", "%A6%EC", "GB18030-2022 8");
 encode("\uFE18", "%A6%ED", "GB18030-2022 9");
 encode("\uFE19", "%A6%F3", "GB18030-2022 10");
 encode("\u9FB4", "%FEY", "GB18030-2022 11");
 encode("\u9FB5", "%FEa", "GB18030-2022 12");
 encode("\u9FB6", "%FEf", "GB18030-2022 13");
 encode("\u9FB7", "%FEg", "GB18030-2022 14");
 encode("\u9FB8", "%FEm", "GB18030-2022 15");
 encode("\u9FB9", "%FE~", "GB18030-2022 16");
 encode("\u9FBA", "%FE%90", "GB18030-2022 17");
 encode("\u9FBB", "%FE%A0", "GB18030-2022 18");
 encode("\uE78D", "%A6%D9", "GB18030-2022 19");
 encode("\uE78E", "%A6%DA", "GB18030-2022 20");
 encode("\uE78F", "%A6%DB", "GB18030-2022 21");
 encode("\uE790", "%A6%DC", "GB18030-2022 22");
 encode("\uE791", "%A6%DD", "GB18030-2022 23");
 encode("\uE792", "%A6%DE", "GB18030-2022 24");
 encode("\uE793", "%A6%DF", "GB18030-2022 25");
 encode("\uE794", "%A6%EC", "GB18030-2022 26");
 encode("\uE795", "%A6%ED", "GB18030-2022 27");
 encode("\uE796", "%A6%F3", "GB18030-2022 28");
 encode("\uE81E", "%FEY", "GB18030-2022 29");
 encode("\uE826", "%FEa", "GB18030-2022 30");
 encode("\uE82B", "%FEf", "GB18030-2022 31");
 encode("\uE82C", "%FEg", "GB18030-2022 32");
 encode("\uE832", "%FEm", "GB18030-2022 33");
 encode("\uE843", "%FE~", "GB18030-2022 34");
 encode("\uE854", "%FE%90", "GB18030-2022 35");
 encode("\uE864", "%FE%A0", "GB18030-2022 36");
</script>
