third_party/blink/web_tests/external/wpt/encoding/legacy-mb-schinese/gbk/gbk-encoder-expected.txt

This is a testharness.js-based test.
[FAIL] gbk encoder: GB18030-2022 19
  assert_equals: expected "%A6%D9" but got "%26%2359277%3B"
[FAIL] gbk encoder: GB18030-2022 20
  assert_equals: expected "%A6%DA" but got "%26%2359278%3B"
[FAIL] gbk encoder: GB18030-2022 21
  assert_equals: expected "%A6%DB" but got "%26%2359279%3B"
[FAIL] gbk encoder: GB18030-2022 22
  assert_equals: expected "%A6%DC" but got "%26%2359280%3B"
[FAIL] gbk encoder: GB18030-2022 23
  assert_equals: expected "%A6%DD" but got "%26%2359281%3B"
[FAIL] gbk encoder: GB18030-2022 24
  assert_equals: expected "%A6%DE" but got "%26%2359282%3B"
[FAIL] gbk encoder: GB18030-2022 25
  assert_equals: expected "%A6%DF" but got "%26%2359283%3B"
[FAIL] gbk encoder: GB18030-2022 26
  assert_equals: expected "%A6%EC" but got "%26%2359284%3B"
[FAIL] gbk encoder: GB18030-2022 27
  assert_equals: expected "%A6%ED" but got "%26%2359285%3B"
[FAIL] gbk encoder: GB18030-2022 28
  assert_equals: expected "%A6%F3" but got "%26%2359286%3B"
[FAIL] gbk encoder: GB18030-2022 29
  assert_equals: expected "%FEY" but got "%26%2359422%3B"
[FAIL] gbk encoder: GB18030-2022 30
  assert_equals: expected "%FEa" but got "%26%2359430%3B"
[FAIL] gbk encoder: GB18030-2022 31
  assert_equals: expected "%FEf" but got "%26%2359435%3B"
[FAIL] gbk encoder: GB18030-2022 32
  assert_equals: expected "%FEg" but got "%26%2359436%3B"
[FAIL] gbk encoder: GB18030-2022 33
  assert_equals: expected "%FEm" but got "%26%2359442%3B"
[FAIL] gbk encoder: GB18030-2022 34
  assert_equals: expected "%FE~" but got "%26%2359459%3B"
[FAIL] gbk encoder: GB18030-2022 35
  assert_equals: expected "%FE%90" but got "%26%2359476%3B"
[FAIL] gbk encoder: GB18030-2022 36
  assert_equals: expected "%FE%A0" but got "%26%2359492%3B"
Harness: the test ran to completion.
